test: ProductsFixture 이름 변경 및 테스트 데이터 수정
